requirements.txt
tensorflow
numpy
streamlit
pandas
Pillow
opencv-python-headless
pytest
seaborn
matplotlib
scikit-learn
